reading_tasks/requirements.txt
newspaper3k
Flask
newsapi-python
cohere

